# api/docs/openapi.yml
openapi: "3.0.3"
info:
  title: PlayBingo API
  version: "1.0"
servers:
  - url: https://playbingo.gg/api
...
components:
  securitySchemes:
    cookieAuthentication:
      description:
        Authentication is handled via session and cookies. Endpoints
        protected by this authentication method are only available to requests
        made by the website, unless they list another form of authorization. The
        value of the cookie indicates the server session of the currently logged
        in user, which is used to identify the user across page loads and
        multiple requests.
      type: apiKey
      in: cookie
      name: connect.sid
    apiKeyAuthentication:
      description: 
        Authentication is handled via a header. Endpoints protected with this
        authentication method are considered "private" and are not available for
        consumption by general consumers of the API. API keys can only be
        granted by site admins. Each API key uniquely represents an API consumer
        who has been granted access to these private endpoints
      type: apiKey
      in: header
      name: PlayBingo-Api-Key
  responses:
    errorBadRequest:
      description:
        "Bad Request: Required parameters, request body, or body contents are
        missing, malformed, or otherwise invalid, and the client should have
        been able to validate them as such."
    errorUnauthorized:
      description:
        "Unauthorized: The request failed to provide identification in a way the API
        server recognizes."
    errorForbidden:
      description:
        "Forbidden: The provided authorization does not have permission to
        perform the requested action"
    errorNotFound:
      description:
        "Not Found: The requested resource was not found"
    errorServer:
      description:
        "Internal Server Error: The server encountered an error unrelated to the
        contents of the request"
  schemas:
...
    game:
      type: object
      description: Represents the high level information for a single bingo game
      required:
        - name
        - slug
        - owners
        - moderators
        - racetimeBeta
      properties:
        name:
          type: string
          description: The full name of the game
        slug:
          type: string
          description: The game's unique, short name used for URLs
        coverImage:
          type: string
          description:
            URL to the game's cover image displayed on the website. May link
            externally
        owners:
          type: array
          description: The owners of the game
          items:
            $ref: "#/components/schemas/user"
        moderators:
          type: array
          description: The moderators of the game
          items:
            $ref: "#/components/schemas/user"
        enableSRLv5:
          type: boolean
          description: If the game is enabled for SRLv5 board generation
        racetimeBeta:
          type: boolean
          description:
            \[TEMPORARY\] Flag indicating if the game has opted in to the
            racetime.gg integration beta program
        racetimeCategory:
          type: string
          description: The slug of the connected racetime category
        racetimeGoal:
          type: string
          description: The name of the connected racetime goal
        difficultyVariantsEnabled:
          type: boolean
          description: If the game has enabled difficulty variants
        difficultyVariants:
          type: array
          description: The game's configured difficulty variants
          items:
            $ref: "#/components/schemas/difficultyVariant"
        difficultyGroups:
          type: number
          description:
            The number of unique difficulty groups used in difficulty
            variant generation
        slugWords:
          type: array
          description:
            Custom slug words (typically nouns) used for room
            identifiers for this game
          items:
            type: string
        useTypedRandom:
          type: boolean
          description:
            If this game should consider goal types during
            generation, mimicking SRLv5's line logic
...
paths:
...
  /games:
    get:
      summary: Get all games
      description: Gets the full list of games.
      tags:
      - games
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/game"
    post:
      summary: Create a new game
      description:
        Creates a new game from the specified data. Requires a login session,
        and can only be called from the website. The user in the login session
        will be made the owner of the game.
      tags:
        - games
      security:
        - cookieAuthentication: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - name
                - slug
              properties:
                name:
                  type: string
                  description: The name of the new game
                slug:
                  type: string
                  description: The slug of the new game
                coverImage:
                  type: string
                  description: Link to the cover image of the new game
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/game"
        400:
          $ref: "#/components/responses/errorBadRequest"
        401:
          $ref: "#/components/responses/errorUnauthorized"
        500:
          $ref: "#/components/responses/errorServer"
  /games/{slug}:
    get:
      summary: Get a single game
      description: Gets data for a single game
      tags:
        - games
      parameters:
        - name: slug
          in: path
          required: true
          description: The unique identifier for the game.
          schema:
            type: string
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/game"
        404:
          $ref: "#/components/responses/errorNotFound"
    post:
      summary: Update a game
      description:
        Update the data for an existing game
      tags:
        - games
      parameters:
        - name: slug
          in: path
          required: true
          description: The unique identifier for the game.
          schema:
            type: string
      security:
        - cookieAuthentication: []
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/game"
        400:
          $ref: "#/components/responses/errorBadRequest"
    delete:
      summary: Delete a game
      description: Deletes the game identified by the provided slug.
      tags:
        - games
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
          description: The unique identifier for the game.
      responses:
        200:
          description: OK
        401:
          $ref: "#/components/responses/errorUnauthorized"
        403:
          $ref: "#/components/responses/errorForbidden"
  /games/{slug}/goals:
    get:
      summary: Get goals for a game
      description:
        Get the list of all goals for a game
      tags:
        - games
      parameters:
        - name: slug
          in: path
          required: true
          description: The unique identifier for the game.
          schema:
            type: string
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/goal"
    post:
      summary: Create a new goal
      description:
        Creates a new goal belonging to the specified game
      tags:
        - games
      parameters:
        - name: slug
          in: path
          required: true
          description: The unique identifier for the game.
          schema:
            type: string
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/goal"
        400:
          $ref: "#/components/responses/errorBadRequest"
  /games/{slug}/eligibleMods:
    get:
      summary: Get eligible moderators
      description: Returns the list of users eligible to be a moderator or owner of the game.
      tags:
          - games
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
          description: The unique identifier for the game.
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/user"
  /games/{slug}/owners:
    get:
      summary: Get game owners
      description: Get the users who are owners for the game.
      tags:
        - games
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
          description: The unique identifier for the game.
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/user"
        404:
          $ref: "#/components/responses/errorNotFound"
    post:
      summary: Add game owners
      description: Adds new owners to the game.
      tags:
        - games
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
          description: The unique identifier for the game.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - users
              properties:
                users:
                  type: array
                  items:
                    type: string
                  description: List of user IDs to add as owners.
      responses:
        200:
          description: OK
        400:
          $ref: "#/components/responses/errorBadRequest"
        401:
          $ref: "#/components/responses/errorUnauthorized"
        403:
          $ref: "#/components/responses/errorForbidden"

    delete:
      summary: Remove game owners
      description:
        Removes the specified owner from the game. A game must always have at
        least one owner, and the last remaining owner cannot be removed with
        this endpoint.
      tags:
        - games
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
          description: The unique identifier for the game.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - user
              properties:
                user:
                  type: string
                  description: The user ID to remove as an owner.
      responses:
        200:
          description: OK
        400:
          $ref: "#/components/responses/errorBadRequest"
        401:
          $ref: "#/components/responses/errorUnauthorized"
        403:
          $ref: "#/components/responses/errorForbidden"
        404:
          $ref: "#/components/responses/errorNotFound"
  /games/{slug}/moderators:
    get:
      summary: Get game moderators
      description: Get the users who are moderators for the game.
      tags:
        - games
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
          description: The unique identifier for the game.
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/user"
        404:
          $ref: "#/components/responses/errorNotFound"
    post:
      summary: Add game moderators
      description: Adds new moderators to the game.
      tags:
        - games
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
          description: The unique identifier for the game.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - users
              properties:
                users:
                  type: array
                  items:
                    type: string
                  description: List of user IDs to add as moderators.
      responses:
        200:
          description: OK
        400:
          $ref: "#/components/responses/errorBadRequest"
        401:
          $ref: "#/components/responses/errorUnauthorized"
        403:
          $ref: "#/components/responses/errorForbidden"

    delete:
      summary: Remove game moderators
      description: Removes the specified moderator from the game.
      tags:
        - games
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
          description: The unique identifier for the game.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - user
              properties:
                user:
                  type: string
                  description: The user ID to remove as a moderator.
      responses:
        200:
          description: OK
        400:
          $ref: "#/components/responses/errorBadRequest"
        401:
          $ref: "#/components/responses/errorUnauthorized"
        403:
          $ref: "#/components/responses/errorForbidden"
        404:
          $ref: "#/components/responses/errorNotFound"
  /games/{slug}/permissions:
    get:
      summary: Get user permissions
      description: Determines the currently logged-in user's permissions for the game.
      tags:
        - games
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
          description: The unique identifier for the game.
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  isOwner:
                    type: boolean
                    description: Whether the user is an owner of the game.
                  canModerate:
                    type: boolean
                    description: Whether the user can moderate the game.
        401:
          $ref: "#/components/responses/errorUnauthorized"
  /games/{slug}/deleteAllGoals:
    delete:
      summary: Delete all game goals
      description: Deletes all goals for the game.
      tags:
        - games
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
          description: The unique identifier for the game.
      responses:
        200:
          description: OK
        400:
          $ref: "#/components/responses/errorBadRequest"
        401:
          $ref: "#/components/responses/errorUnauthorized"
        403:
          $ref: "#/components/responses/errorForbidden"
        500:
          $ref: "#/components/responses/errorServer"
  /games/{slug}/favorite:
    post:
      summary: Favorite a game
      description:
        Adds the game to the favorites of the currently logged-in user.
      tags:
        - games
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
          description: The unique identifier for the game.
      responses:
        200:
          description: OK
        401:
          $ref: "#/components/responses/errorUnauthorized"
    delete:
      summary: Unfavorite a game
      description:
        Removes the game from the favorites of the currently logged-in user.
      tags:
        - games
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
          description: The unique identifier for the game.
      responses:
        200:
          description: OK
        401:
          $ref: "#/components/responses/errorUnauthorized"
  /games/{slug}/difficultyVariants:
    post:
      summary: Create a difficulty variant
      description: Creates a new difficulty variant for the game.
      tags:
        - games
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
          description: The unique identifier for the game.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - goalAmounts
              properties:
                name:
                  type: string
                  description: The name of the difficulty variant.
                goalAmounts:
                  type: array
                  items:
                    type: number
                  description: The number of goals in each difficulty group.
      responses:
        200:
          description: OK
        400:
          $ref: "#/components/responses/errorBadRequest"
        401:
          $ref: "#/components/responses/errorUnauthorized"
        403:
          $ref: "#/components/responses/errorForbidden"
  /games/{slug}/difficultyVariants/{id}:
    post:
      summary: Update a difficulty variant
      description: Updates the specified difficulty variant for the game.
      tags:
        - games
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
          description: The unique identifier for the game.
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: The unique identifier for the difficulty variant.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - goalAmounts
              properties:
                name:
                  type: string
                  description: The name of the difficulty variant.
                goalAmounts:
                  type: array
                  items:
                    type: number
                  description: The number of goals in each difficulty group.
      responses:
        200:
          description: OK
        400:
          $ref: "#/components/responses/errorBadRequest"
        401:
          $ref: "#/components/responses/errorUnauthorized"
        403:
          $ref: "#/components/responses/errorForbidden"

    delete:
      summary: Delete a difficulty variant
      description: Deletes the specified difficulty variant from the game.
      tags:
        - games
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
          description: The unique identifier for the game.
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: The unique identifier for the difficulty variant.
      responses:
        200:
          description: OK
        401:
          $ref: "#/components/responses/errorUnauthorized"
        403:
          $ref: "#/components/responses/errorForbidden"
  /games/{slug}/categories:
    get:
      summary: Get game categories
      description: Retrieves the list of categories for the game.
      tags:
        - games
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
          description: The unique identifier for the game.
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                      description: The unique identifier for the category.
                    name:
                      type: string
                      description: The name of the category.
                    max:
                      type: integer
                      description: The maximum number of goals allowed in the category.
                    goalCount:
                      type: integer
                      description: The number of goals in the category.
        401:
          $ref: "#/components/responses/errorUnauthorized"

    post:
      summary: Create a category
      description: Creates a new category for the game.
      tags:
        - games
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
          description: The unique identifier for the game.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  description: The name of the category.
                max:
                  type: integer
                  description: The maximum number of goals allowed in the category.
      responses:
        200:
          description: OK
        400:
          $ref: "#/components/responses/errorBadRequest"
        401:
          $ref: "#/components/responses/errorUnauthorized"
        403:
          $ref: "#/components/responses/errorForbidden"